refactor: common teams query params in requests
