+++ src/rubrix/server/tasks/token_classification/api/api.py
@@ -16,6 +16,8 @@
 import itertools
 from fastapi import APIRouter, Depends, Query, Security
 from fastapi.responses import StreamingResponse
+
+from rubrix.server.commons.api import TeamsQueryParams
 from rubrix.server.datasets.model import CreationDatasetRequest
 from rubrix.server.datasets.service import DatasetsService, create_dataset_service
 from rubrix.server.security import auth
@@ -24,7 +26,6 @@ from rubrix.server.tasks.commons import (
     BulkResponse,
     PaginationParams,
     TaskType,
-    TeamsQueryParams,
 )
 from rubrix.server.tasks.commons.helpers import takeuntil
 from rubrix.server.tasks.token_classification.api.model import (
